+++ SpatialGDK/Source/SpatialGDK/Private/Interop/SpatialStaticComponentView.cpp
@@ -14,7 +14,6 @@
 #include "Schema/RPCPayload.h"
 #include "Schema/ServerEndpoint.h"
 #include "Schema/ServerRPCEndpointLegacy.h"
-#include "Schema/Singleton.h"
 #include "Schema/SpatialDebugging.h"
 #include "Schema/SpawnData.h"
 #include "Schema/UnrealMetadata.h"
@@ -49,6 +48,13 @@ bool USpatialStaticComponentView::HasComponent(Worker_EntityId EntityId, Worker_
 
 void USpatialStaticComponentView::OnAddComponent(const Worker_AddComponentOp& Op)
 {
+	// With dynamic components enabled, it's possible to get duplicate AddComponent ops which need handling idempotently.
+	// For the sake of efficiency we just exit early here.
+	if (HasComponent(Op.entity_id, Op.data.component_id))
+	{
+		return;
+	}
+
 	TUniquePtr<SpatialGDK::Component> Data;
 	switch (Op.data.component_id)
 	{
@@ -69,9 +75,6 @@ void USpatialStaticComponentView::OnAddComponent(const Worker_AddComponentOp& Op
 		break;
 	case SpatialConstants::SPAWN_DATA_COMPONENT_ID:
 		Data = MakeUnique<SpatialGDK::SpawnData>(Op.data);
-		break;
-	case SpatialConstants::SINGLETON_COMPONENT_ID:
-		Data = MakeUnique<SpatialGDK::Singleton>(Op.data);
 		break;
 	case SpatialConstants::UNREAL_METADATA_COMPONENT_ID:
 		Data = MakeUnique<SpatialGDK::UnrealMetadata>(Op.data);